read_char: Handle non-tty terminals explicitly
`Key::Unknown` is produced only when:
- A keycode is not recognized by the `key_from_key_code` function under Windows
- `read_key` is invoked on a non-tty terminal

The two places in the code which explicitly handled `Key::Unknown` were
treating it as a proxy for an "unattended terminal" check. Since this is
only 100% valid to do when running under *nix I've replaced these
handlers with an explicit `!self.is_tty` check instead, which is a common
pattern in the Term code.

`Key::Unknown` is now assumed to represent an unprintable / control key
and skipped when being evaluated.