Moved various reading functions to TermTarget
Its better to have only one place that checks if the TermTarget is a ReadWritePair instead of having them everywhere in the code. As the way input is read is determined by the TermTarget it made sense to move it there. I still need to implement the functionality for a ReadWritePairSource though