Update read_line behaviour to read_line_initial_text
`term.read_line()` is not using the declared term to request the input. It relies on stdin.
This is not the case of `term.read_line_initial_text("")`, which reads key one by one using `self.read_key()`.

This commits corrects this behaviour, and removes divergence between the two methods.