Fix potential creation of an unaligned reference
The `name_info_bytes` Vec is only guaranteed to have an alignment of 1
(`std::mem::align_of::<u8>()`), however a raw pointer to it gets turned into a `&FILE_NAME_INFO`,
which has an alignment requirement of 4.
Creating unaligned references is immediate undefined behavior.

This bug was copied from `atty` and is tracked there as RUSTSEC-2021-0145 [1]. The fix in this
commit is taken from `atty`'s fork `is-terminal` [2], which was created since `atty` is
unmaintained.

Note that this code only gets compiled in on Windows and Windows' system allocator guarantees a
large enough alignment. Thus, in practice the pointer could only be unaligned if a custom global
allocator is used.

[1] https://rustsec.org/advisories/RUSTSEC-2021-0145.html
[2] https://github.com/sunfishcode/is-terminal/commit/0c8d91d5a3b77d510971718a4210e491d9d60366